project-2 cambios

+++ list-edit-router.js
@@ -13,19 +13,33 @@ listEditRouter.post("/crear", (req, res) => {
 
 listEditRouter.delete("/borrar/:id", (req, res) => {
   const id = parseInt(req.params.id);
-  funcionTareasEditar.eliminarTarea(id);
-  res.json({
-    message: `la tarea con id ${id} ha sido eliminada`,
-  });
+  const tareaExistente = funcionTareasEditar.buscarTareaPorId(id);
+  if (!tareaExistente) {
+    res.status(404).json({
+      status: 404,
+      message: `No puedes borrar la tarea con id ${id} porque no existe`,
+    });
+  } else {
+    funcionTareasEditar.eliminarTarea(id);
+    res.json({
+      message: `la tarea con id ${id} ha sido eliminada`,
+    });
+  }
 });
 
 listEditRouter.put("/actualizar/:id", (req, res) => {
   const id = parseInt(req.params.id);
-  const { nuevoCompletada } = req.body;
-  if (nuevoCompletada !== undefined) {
-    funcionTareasEditar.actualizarTarea(id, nuevoCompletada);
+  const siExisteTarea = funcionTareasEditar.buscarTareaPorId(id);
+  if (!siExisteTarea) {
+    res.status(404).json({
+      status: 404,
+      message: `No se encontro la tarea con ID ${id}, intenta nuevamente`,
+    });
+  } else {
+    funcionTareasEditar.actualizarTarea(id, req.body);
     res.json({
-      message: `la tarea ${id} ha sido actualizada a ${nuevoCompletada}`,
+      status: 200,
+      message: `la tarea con id ${req.params.id} fue editada exitosamente`,
     });
   }
 });
